Add AgentAwsUser, refactor into aws dir
