+++ typescript/infra/src/agents/aws/index.ts
@@ -1,0 +1,3 @@
+export { AgentAwsKey } from './key';
+export { AgentAwsUser } from './user';
+export { ValidatorAgentAwsUser } from './validator-user';
