+++ typescript/infra/src/agents/aws/key.ts
@@ -1,6 +1,6 @@
 import { ChainName } from '@abacus-network/sdk';
-import { getSecretAwsCredentials, KEY_ROLE_ENUM } from '../agents';
-import { AgentConfig } from '../../src/config/agent';
+import { getSecretAwsCredentials, KEY_ROLE_ENUM } from '..';
+import { AgentConfig } from '../../config/agent';
 import {
   CreateAliasCommand,
   CreateKeyCommand,
@@ -13,21 +13,8 @@ import {
   OriginType,
   UpdateAliasCommand,
 } from '@aws-sdk/client-kms';
-import {
-  IAMClient,
-  CreateAccessKeyCommand,
-  CreateUserCommand,
-  ListUsersCommand,
-} from '@aws-sdk/client-iam';
-import {
-  S3Client,
-  CreateBucketCommand,
-  ListBucketsCommand,
-  PutBucketPolicyCommand,
-} from '@aws-sdk/client-s3';
-import { getEthereumAddress } from '../utils/utils';
-import { AgentKey } from './agent';
-import { gcpSecretExists, setGCPSecret } from '../utils/gcloud';
+import { getEthereumAddress } from '../../utils/utils';
+import { AgentKey } from '../agent';
 
 interface UnfetchedKey {
   fetched: false;
@@ -39,176 +26,6 @@ interface FetchedKey {
 }
 
 type RemoteKey = UnfetchedKey | FetchedKey;
-
-export class ValidatorAgentAwsUser<Networks extends ChainName> {
-  private adminIamClient: IAMClient;
-  private adminS3Client: S3Client;
-
-  private _arn: string | undefined;
-
-  constructor(
-    public readonly environment: string,
-    public readonly chainName: Networks,
-    public readonly index: number,
-    public readonly region: string,
-    public readonly bucket: string,
-  ) {
-    this.adminIamClient = new IAMClient({ region });
-    this.adminS3Client = new S3Client({ region });
-  }
-
-  // Creates the AWS user if it doesn't exist.
-  // Gets API access keys and saves them in GCP secret manager if they do not already exist.
-  // Populates `this._arn` with the ARN of the user.
-  async createIfNotExists() {
-    const cmd = new ListUsersCommand({});
-    const result = await this.adminIamClient.send(cmd);
-    const match = result.Users?.find((user) => user.UserName === this.userName);
-    if (match) {
-      this._arn = match?.Arn;
-    } else {
-      this._arn = await this.create();
-    }
-    if (!(await this.accessKeysExist())) {
-      await this.createAndSaveAccessKey();
-    }
-  }
-
-  // Creates the AWS user
-  async create() {
-    const cmd = new CreateUserCommand({
-      UserName: this.userName,
-      Tags: this.awsTags,
-    });
-    const response = await this.adminIamClient.send(cmd);
-    if (!response.User) {
-      throw Error('Expected User');
-    }
-    return response.User.Arn;
-  }
-
-  async accessKeysExist(): Promise<boolean> {
-    const accessKeyIdExists = await gcpSecretExists(this.accessKeyIdSecretName);
-    const secretAccessKeyExists = await gcpSecretExists(
-      this.secretAccessKeySecretName,
-    );
-    return accessKeyIdExists && secretAccessKeyExists;
-  }
-
-  async createAndSaveAccessKey() {
-    const cmd = new CreateAccessKeyCommand({
-      UserName: this.userName,
-    });
-    const { AccessKey: accessKey } = await this.adminIamClient.send(cmd);
-    if (!accessKey || !accessKey.AccessKeyId || !accessKey.SecretAccessKey) {
-      throw Error('Expected fully defined access key');
-    }
-    await setGCPSecret(
-      this.accessKeyIdSecretName,
-      accessKey.AccessKeyId,
-      this.tags,
-    );
-    await setGCPSecret(
-      this.secretAccessKeySecretName,
-      accessKey.SecretAccessKey,
-      this.tags,
-    );
-  }
-
-  async createBucketIfNotExists() {
-    if (!(await this.bucketExists())) {
-      await this.createBucket();
-    }
-    await this.putBucketAccessPolicy();
-  }
-
-  async bucketExists(): Promise<boolean> {
-    const cmd = new ListBucketsCommand({});
-    const result = await this.adminS3Client.send(cmd);
-    return (
-      result.Buckets?.find((bucket) => bucket.Name === this.bucket) !==
-      undefined
-    );
-  }
-
-  async createBucket() {
-    const cmd = new CreateBucketCommand({
-      Bucket: this.bucket,
-    });
-    await this.adminS3Client.send(cmd);
-  }
-
-  async putBucketAccessPolicy() {
-    const policy = {
-      Statement: [
-        // Make the bucket publicly readable
-        {
-          Effect: 'Allow',
-          Principal: '*',
-          Action: ['s3:GetObject', 's3:ListBucket'],
-          Resource: [
-            `arn:aws:s3:::${this.bucket}`,
-            `arn:aws:s3:::${this.bucket}/*`,
-          ],
-        },
-        // Allow the user to modify objects
-        {
-          Effect: 'Allow',
-          Principal: {
-            AWS: this.arn,
-          },
-          Action: ['s3:DeleteObject', 's3:PutObject'],
-          Resource: `arn:aws:s3:::${this.bucket}/*`,
-        },
-      ],
-    };
-    const cmd = new PutBucketPolicyCommand({
-      Bucket: this.bucket,
-      Policy: JSON.stringify(policy),
-    });
-    await this.adminS3Client.send(cmd);
-  }
-
-  get awsTags() {
-    const tags = this.tags;
-    return Object.keys(tags).map((key) => ({
-      Key: key,
-      Value: tags[key],
-    }));
-  }
-
-  get tags(): Record<string, string> {
-    return {
-      environment: this.environment,
-      role: this.role,
-      chain: this.chainName,
-      index: this.index!.toString(),
-    };
-  }
-
-  get role() {
-    return KEY_ROLE_ENUM.Validator;
-  }
-
-  get userName() {
-    return `abacus-${this.environment}-${this.chainName}-${this.role}-${this.index}`;
-  }
-
-  get accessKeyIdSecretName() {
-    return `${this.userName}-aws-access-key-id`;
-  }
-
-  get secretAccessKeySecretName() {
-    return `${this.userName}-aws-secret-access-key`;
-  }
-
-  get arn(): string {
-    if (!this._arn) {
-      throw Error('ARN is undefined');
-    }
-    return this._arn;
-  }
-}
 
 export class AgentAwsKey<Networks extends ChainName> extends AgentKey {
   private environment: string;
